07.04.22

+++ simulink/modbus-connector/modbus_write.m
@@ -72,4 +72,4 @@ function ModBusTCP = openConnection(ipaddress, port)
   ModBusTCP.ByteOrder='big-endian';
 
 function writeModBus(ModBusTCP, target, registry, value)
-  write(ModBusTCP, target, registry, value, 'double')
+  write(ModBusTCP, target, registry, value, 'single')
